+++ apps/crm/crm/src/app/crm/pages/offers/shell/contract-form/contract-form.component.html
@@ -1,1 +1,4 @@
-<negotiation-form [form]="form"> </negotiation-form>
+<negotiation-form
+  [form]="form" [title]="title" [currency]="currency"
+  (save)="save()"
+> </negotiation-form>
