+++ apps/crm/crm/src/app/crm/pages/offers/shell/contract-form/contract-form.module.ts
@@ -15,7 +15,6 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
 // @blockframes
 import { AvailsFilterModule } from '@blockframes/contract/avails/filter/filter.module';
-import { FormTableModule } from "@blockframes/ui/form/table/form-table.module";
 import { TableFilterModule } from "@blockframes/ui/list/table-filter/table-filter.module";
 import { LanguagesFormModule } from "@blockframes/movie/form/languages/languages.module";
 import { NegotiationModule } from '@blockframes/contract/negotiation'
@@ -32,7 +31,6 @@ import { ToGroupLabelPipeModule } from '@blockframes/utils/pipes/group-label.pip
     RouterModule.forChild([{ path: '', component: ContractFormComponent }]),
     ReactiveFormsModule,
     AvailsFilterModule,
-    FormTableModule,
     TableFilterModule,
     ToLabelModule,
     VersionPipeModule,
