+++ libs/contract/src/lib/negotiation/form/negotiation-form.component.html
@@ -1,9 +1,9 @@
-<form [formGroup]="form" (ngSubmit)="save()">
+<form [formGroup]="form" (ngSubmit)="onSave()">
   <header>
     <h2>{{ title?.title.international }}</h2>
     <mat-form-field appearance="outline">
       <mat-label>Title price</mat-label>
-      <input type="number" matInput formControlName="price" min="0"/>
+      <input type="number" matInput formControlName="price" min="0" />
       <mat-icon matPrefix *ngIf="currency" [svgIcon]="currency"></mat-icon>
     </mat-form-field>
   </header>
