+++ libs/contract/src/lib/negotiation/form/negotiation.module.ts
@@ -5,6 +5,7 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { AvailsFilterModule } from '@blockframes/contract/avails/filter/filter.module';
 import { LanguagesFormModule } from '@blockframes/movie/form/languages/languages.module';
 import { ToLabelModule, VersionPipeModule, JoinPipeModule, ToGroupLabelPipeModule } from '@blockframes/utils/pipes';
+import { FormTableModule } from "@blockframes/ui/form/table/form-table.module";
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
@@ -16,6 +17,7 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
     ReactiveFormsModule,
     AvailsFilterModule,
     MatProgressSpinnerModule,
+    FormTableModule,
     MatFormFieldModule,
     MatIconModule,
     ToLabelModule,
